update to new modevlib

+++ html/modevlib/rest/Rest.js
@@ -38,7 +38,7 @@ Rest.send=function*(ajaxParam){
 	if (ajaxParam.dataType===undefined) ajaxParam.dataType="json";
 	if (ajaxParam.error===undefined){
 		ajaxParam.error=function(errorData){
-			callback(new Exception("Error while calling "+ajaxParam.url+": "+errorData));
+			callback(new Exception("Error while calling "+ajaxParam.url, errorData));
 		};
 	}//endif
 	if (typeof(ajaxParam.data)!="string") ajaxParam.data=CNV.Object2JSON(ajaxParam.data);
@@ -77,7 +77,7 @@ Rest.send=function*(ajaxParam){
 
 	request.onreadystatechange = function(){
 		if (request.readyState == 4){
-			if (request.status == 200){
+			if ([200, 201].contains(request.status)){
 				var response = request.responseText;
 				if (ajaxParam.dataType == 'json'){
 					response = CNV.JSON2Object(response);
@@ -89,7 +89,7 @@ Rest.send=function*(ajaxParam){
 			} else if (request.isTimeout){
 				callback(new Exception("Error while calling " + ajaxParam.url, Exception.TIMEOUT));
 			} else {
-				ajaxParam.error("Bad response: " + CNV.String2Quote(request.responseText));
+				ajaxParam.error(new Exception("Bad response ("+request.status+")", CNV.String2Quote(request.responseText)));
 			}//endif
 		} else if (request.readyState == 3){
 			//RESPONSE IS ARRIVING, DISABLE TIMEOUT
@@ -118,8 +118,6 @@ Rest.send=function*(ajaxParam){
 		}catch(e){
 			//IGNORE
 		}//try
-		//WE DO NOT CALL THE CALLBACK BECAUSE THE THREAD WILL BE CALLING INTERRUPT SHORTLY
-		//callback(new Exception("Aborted"));
 	};
 
 	if (ajaxParam.timeout!==undefined){
@@ -134,6 +132,7 @@ Rest.send=function*(ajaxParam){
 		);
 	}
 
+	request.name="XMLHttpRequest";  //FOR DEBUGGING
 	request.send(ajaxParam.data);
 	yield (Thread.suspend((ajaxParam.doNotKill) ? undefined : request));
 };//method
@@ -154,11 +153,14 @@ Rest.post=function(ajaxParam){
 	return Rest.send(ajaxParam);
 };//method
 
-Rest["delete"]=function(ajaxParam){
-//	Log.warning("DISABLED DELETE OF "+ajaxParam.url);
-//	yield (null);
-	ajaxParam.type="DELETE";
-	return Rest.send(ajaxParam);
+Rest["delete"]=function*(ajaxParam){
+	if (ajaxParam.url.indexOf("org_chart")>0){
+		ajaxParam.type="DELETE";
+		yield (Rest.send(ajaxParam));
+	}else{
+		Log.warning("DISABLED DELETE OF "+ajaxParam.url);
+		yield (null);
+	}
 };//method
